Add: - change currency button to front page - moved all necessary things to StartFragment
